Add converter mock

+++ mocks/router/src/main/scala/net/golikov/Transfer.scala
@@ -1,9 +1,9 @@
 package net.golikov
 
-import io.circe.{Decoder, Encoder}
+import io.circe.{ Decoder, Encoder }
 import io.circe.generic.semiauto._
 
-case class Transfer(id: Option[BigDecimal], content: Array[Byte])
+case class Transfer(id: Option[Long], content: Array[Byte])
 
 object Transfer {
   implicit val encoder: Encoder[Transfer] = deriveEncoder
